ABI: Encode dates as UNIX milliseconds (#1808)
Reuse Partout JSONEncoder/JSONDecoder in both ABI and CommonProviders.
Might softly break serialized providers' `lastUpdate`, but profiles are
unaffected. ABI only uses `Date` in `OriginalPurchase.purchaseDate`.
Make providers `Timestamp` consistent with Partout encoding.

See https://github.com/partout-io/partout/pull/390

+++ app-cross/Sources/CommonProvidersCore/ProviderScriptingAPI.swift
@@ -20,9 +20,9 @@ public protocol ProviderScriptingAPI: Sendable {
 
     func jsonToBase64(object: Any) -> String?
 
-    func timestampFromISO(isoString: String) -> Int
+    func timestampFromISO(isoString: String) -> Int64
 
-    func timestampToISO(timestamp: Int) -> String
+    func timestampToISO(timestamp: Int64) -> String
 
     func ipV4ToBase64(ip: String) -> String?
